teach-comparison

+++ 1-amaliy topshiriq.cs
@@ -5,7 +5,7 @@ partial class Program
 	static void Main()
 	{
 		Console.WriteLine("Doira radiusini kiriting: ");
-		double radius = Convert.ToInt32(Console.ReadLine());
+		double radius =Convert.ToDouble(Console.ReadLine());
 		double pi = 3.14;
 		double doirayuzi = pi * radius * radius;
 		double doiraperimetri = 2 * pi * radius;
